Move DragHandle to the parent DataGrid component
