+++ src/common/types.ts
@@ -113,7 +113,6 @@ export interface CellRendererProps<TRow, TSummaryRow = unknown> extends Omit<Rea
   isCopied: boolean;
   isDraggedOver: boolean;
   eventBus: EventBus;
-  enableCellDragAndDrop: boolean;
   onRowClick?: (rowIdx: number, row: TRow, column: CalculatedColumn<TRow, TSummaryRow>) => void;
 }
 
@@ -128,10 +127,10 @@ export interface RowRendererProps<TRow, TSummaryRow = unknown> extends Omit<Reac
   draggedOverCellIdx?: number;
   isRowSelected: boolean;
   eventBus: EventBus;
+  top: number;
   onRowClick?: (rowIdx: number, row: TRow, column: CalculatedColumn<TRow, TSummaryRow>) => void;
   rowClass?: (row: TRow) => string | undefined;
-  top: number;
-  enableCellDragAndDrop: boolean;
+  setDraggedOverRowIdx: (overRowIdx: number) => void;
 }
 
 export interface FilterRendererProps<TRow, TFilterValue = unknown, TSummaryRow = unknown> {
